requirements.txt
aiogram>=3.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0
python-dateutil>=2.8.2 